Refactored reasoner integration tests
