+++ .circleci/config.yml
@@ -47,13 +47,24 @@ jobs:
       - run: bazel test //test-integration/graql/internal/... --test_output=errors
     # - run: bazel test //test-integration/graql/shell/... --test_output=errors --TODO re add shell tests as soon as fixed
 
-  test-integration-reasoner:
+  test-integration-reasoner-general:
     machine: true
     working_directory: ~/grakn
     steps:
     - checkout
     - *install_bazel
-    - run: bazel test //test-integration/graql/reasoner/... --test_output=errors
+    - run: bazel test //test-integration/graql/reasoner:all --test_output=errors
+
+  test-integration-reasoner-components:
+    machine: true
+    working_directory: ~/grakn
+    steps:
+    - checkout
+    - *install_bazel
+    - run: bazel test //test-integration/graql/reasoner/atomic/... --test_output=errors
+    - run: bazel test //test-integration/graql/reasoner/benchmark/... --test_output=errors
+    - run: bazel test //test-integration/graql/reasoner/cache/... --test_output=errors
+    - run: bazel test //test-integration/graql/reasoner/query/... --test_output=errors
 
   test-integration-analytics:
     machine: true
@@ -196,7 +207,8 @@ workflows:
               ignore: /.*/
       - test-unit
       - test-integration
-      - test-integration-reasoner
+      - test-integration-reasoner-general
+      - test-integration-reasoner-components
       - test-integration-analytics
       - test-end-to-end
       - client-nodejs-e2e
