+++ test-integration/graql/internal/BUILD
@@ -3,7 +3,7 @@ java_test(
     test_class = "grakn.core.graql.internal.IsaExplicitIT",
     srcs = ["IsaExplicitIT.java"],
     deps = [
-        "//test-integration/rule:concurrent-grakn-server",
+        "//test-integration/rule:grakn-test-server",
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
         "//dependencies/maven/artifacts/com/google/guava:guava",
         "//server"],
