+++ test-integration/graql/reasoner/BUILD
@@ -2,9 +2,14 @@ java_test(
     name = "reasoning-it",
     size = "enormous",
     srcs = ["ReasoningIT.java"],
-    classpath_resources = [
-        "//test-integration/resources:logback-test",
-        "//test-integration/resources:cassandra-embedded",
+    test_class = "grakn.core.graql.reasoner.ReasoningIT",
+    deps = [
+        "//dependencies/maven/artifacts/com/google/guava",
+        "//dependencies/maven/artifacts/commons-collections",
+        "//dependencies/maven/artifacts/org/apache/commons:commons-math3",
+        "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
+        "//server",
+        "//test-integration/rule:grakn-test-server",
     ],
     resources = [
         "//test-integration/graql/reasoner/stubs:appending-rps-test",
@@ -46,70 +51,9 @@ java_test(
         "//test-integration/graql/reasoner/stubs:type-derivation-relations-with-direct",
         "//test-integration/graql/reasoner/stubs:type-derivation-with-direct",
     ],
-    test_class = "grakn.core.graql.reasoner.ReasoningIT",
-    deps = [
-        "//dependencies/maven/artifacts/com/google/guava",
-        "//dependencies/maven/artifacts/commons-collections",
-        "//dependencies/maven/artifacts/org/apache/commons:commons-math3",
-        "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
-        "//server",
-        "//test-integration/rule:concurrent-grakn-server",
+    classpath_resources = [
+        "//test-integration/resources:logback-test",
+        "//test-integration/resources:cassandra-embedded",
     ],
 )
 
-java_test(
-    name = "query-validity-it",
-    size = "medium",
-    srcs = ["QueryValidityIT.java"],
-    classpath_resources = ["//test-integration/resources:logback-test"],
-    resources = ["//test-integration/graql/reasoner/resources:rule-applicability"],
-    test_class = "grakn.core.graql.reasoner.QueryValidityIT",
-    deps = [
-        "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
-        "//server",
-        "//test-integration/rule:concurrent-grakn-server",
-    ],
-)
-
-java_test(
-    name = "resolution-plan-it",
-    size = "medium",
-    srcs = ["ResolutionPlanIT.java"],
-    classpath_resources = ["//test-integration/resources:logback-test"],
-    resources = ["//test-integration/graql/reasoner/resources:resolution-plan"],
-    test_class = "grakn.core.graql.reasoner.ResolutionPlanIT",
-    deps = [
-        "//dependencies/maven/artifacts/com/google/guava",
-        "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
-        "//server",
-        "//test-integration/rule:concurrent-grakn-server",
-    ],
-)
-
-java_test(
-    name = "query-it",
-    size = "medium",
-    srcs = ["QueryIT.java"],
-    classpath_resources = ["//test-integration/resources:logback-test"],
-    test_class = "grakn.core.graql.reasoner.QueryIT",
-    deps = [
-        "//server",
-        "//test-integration/graql/reasoner/graph:geo-graph",
-        "//test-integration/rule:concurrent-grakn-server",
-    ],
-)
-
-java_test(
-    name = "ontological-query-it",
-    size = "large",
-    srcs = ["OntologicalQueryIT.java"],
-    classpath_resources = ["//test-integration/resources:logback-test"],
-    data = ["//test-integration/graql/reasoner/resources:rule-applicability"],
-    test_class = "grakn.core.graql.reasoner.OntologicalQueryIT",
-    deps = [
-        "//dependencies/maven/artifacts/com/google/guava",
-        "//server",
-        "//test-integration/rule:concurrent-grakn-server",
-        "//test-integration/util:graql-test-util",
-    ],
-)
