+++ test-integration/graql/reasoner/benchmark/BUILD
@@ -5,7 +5,7 @@ java_test(
     deps = [
             "//server",
             "//client-java",
-            "//test-integration/rule:concurrent-grakn-server",
+            "//test-integration/rule:grakn-test-server",
             "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
             "//dependencies/maven/artifacts/org/slf4j:slf4j-api",
             "//dependencies/maven/artifacts/com/google/guava:guava",
@@ -25,7 +25,7 @@ java_test(
     deps = [
         "//server",
         "//client-java",
-        "//test-integration/rule:concurrent-grakn-server",
+        "//test-integration/rule:grakn-test-server",
         "//dependencies/maven/artifacts/com/google/guava:guava",
         "//test-integration/graql/reasoner/graph:linear-transitivity-matrix-graph",
         "//test-integration/graql/reasoner/graph:transitivity-chain-graph",
