+++ test-integration/graql/reasoner/query/AtomicQueryUnificationIT.java
@@ -44,6 +44,10 @@ import grakn.core.graql.internal.reasoner.query.ReasonerQueries;
 import grakn.core.graql.internal.reasoner.query.ReasonerQueryEquivalence;
 import grakn.core.graql.internal.reasoner.unifier.MultiUnifierImpl;
 import grakn.core.graql.internal.reasoner.unifier.UnifierType;
+import grakn.core.graql.reasoner.query.pattern.RelationPattern;
+import grakn.core.graql.reasoner.query.pattern.ResourcePattern;
+import grakn.core.graql.reasoner.query.pattern.TestQueryPattern;
+import grakn.core.graql.reasoner.query.pattern.TypePattern;
 import grakn.core.server.Session;
 import grakn.core.server.Transaction;
 import grakn.core.server.session.TransactionImpl;
@@ -66,9 +70,9 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import static grakn.core.graql.Graql.var;
-import static grakn.core.graql.reasoner.query.TestQueryPattern.subList;
-import static grakn.core.graql.reasoner.query.TestQueryPattern.subListExcluding;
-import static grakn.core.graql.reasoner.query.TestQueryPattern.subListExcludingElements;
+import static grakn.core.graql.reasoner.query.pattern.TestQueryPattern.subList;
+import static grakn.core.graql.reasoner.query.pattern.TestQueryPattern.subListExcluding;
+import static grakn.core.graql.reasoner.query.pattern.TestQueryPattern.subListExcludingElements;
 import static java.util.stream.Collectors.toSet;
 import static org.apache.commons.collections.CollectionUtils.isEqualCollection;
 import static org.junit.Assert.assertEquals;
