+++ test-integration/graql/reasoner/query/QueryIT.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.graql.reasoner;
+package grakn.core.graql.reasoner.query;
 
 import grakn.core.graql.concept.Concept;
 import grakn.core.graql.Graql;
