+++ test-integration/graql/shell/BUILD
@@ -6,7 +6,7 @@ java_test(
         "//server",
         "//console",
         "//commons",
-        "//test-integration/rule:concurrent-grakn-server",
+        "//test-integration/rule:grakn-test-server",
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
         "//dependencies/maven/artifacts/com/google/guava:guava",
         "//dependencies/maven/artifacts/org/sharegov:mjson",
