+++ test-integration/server/kb/BUILD
@@ -5,7 +5,7 @@ java_test(
      deps = [
          "//server",
          "//commons",
-         "//test-integration/rule:concurrent-grakn-server",
+         "//test-integration/rule:grakn-test-server",
          "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library"
      ],
      classpath_resources = ["//test-integration/resources:logback-test"],
@@ -18,7 +18,7 @@ java_test(
   test_class = "grakn.core.server.kb.ValidateGlobalRulesIT",
   deps = [
       "//server",
-      "//test-integration/rule:concurrent-grakn-server"
+      "//test-integration/rule:grakn-test-server"
   ],
   size = "medium",
   classpath_resources = ["//test-integration/resources:logback-test"]
@@ -31,7 +31,7 @@ java_test(
   deps = [
       "//server",
       "//commons",
-      "//test-integration/rule:concurrent-grakn-server",
+      "//test-integration/rule:grakn-test-server",
       "//dependencies/maven/artifacts/org/apache/tinkerpop:gremlin-core",
       "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
   ],
